WebApp/requirements.txt
flask[async]
python-dotenv
aiohttp